Refactor(SwiftGir): Move Codables from Property Wrappers to DynamicNodeDecoding, DynamicNodeEncoding Protocols
This is done, because the Property Wrappers can't be used with collections, yet.
This all of this should probably replaced with a solution, supporting Property Wrappers on collections and including the ability to serialize into a binary tree, to remove the dependency on SWXMLHash.
I should build that upon GNOME's libxml2.